+++ app/about/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Container from "@/components/container";
-import Title from "@/components/title";
+import Container from '@/components/container';
+import Title from '@/components/title';
 
 const AboutPage = () => {
 	return (
@@ -15,7 +15,7 @@ const AboutPage = () => {
 					<div className="text-neutral-500 font-light">
 						updated April 17th 2024
 					</div>
-					<div className="flex flex-wrap justify-between w-full gap-2 border-y border-foreground/25 py-2">
+					<div className="flex text-neutral-600 flex-wrap justify-between w-full gap-2 border-y border-foreground/25 py-2">
 						<p
 							className="hover:underline cursor-pointer"
 							onClick={() => {
@@ -71,6 +71,16 @@ const AboutPage = () => {
 							{`Analyzing the playlists was a natural addition to the project. The
 							analyzing features are still under construction. Don't know yet how
 							I'll expand this feature set.`}
+						</p>
+					</div>
+					<div className="flex text-neutral-600 justify-center text-center w-full border-t border-foreground/25 pt-2">
+						<p
+							className="hover:underline cursor-pointer"
+							onClick={() => {
+								window.open("https://www.freepik.com/", "_blank");
+							}}
+						>
+							Website favicon by Freepik
 						</p>
 					</div>
 				</div>