+++ app/blocked/page.tsx
@@ -1,5 +1,5 @@
-import Container from '@/components/container';
-import Title from '@/components/title';
+import Container from "@/components/container";
+import Title from "@/components/title";
 
 const BlockedPage = () => {
 	return (
@@ -7,10 +7,10 @@ const BlockedPage = () => {
 			<div className="flex text-center flex-col gap-12 w-full justify-center items-center">
 				<Title />
 				<div>
-					<h1 className="text-2xl text-rose-500">{`YOU'VE BEEN RATELIMITED`}</h1>
+					<h1 className="text-2xl text-rose-500">{`TOO MANY REQUESTS`}</h1>
 				</div>
-				<div className="flex flex-col gap-6 w-full max-w-[800px] bg-secondary p-4 rounded-xl">
-					<p className="font-light text-neutral-500">{`Slow down a bit! Bandwidth doesn't grow on trees.`}</p>
+				<div className="flex flex-col gap-6 w-full max-w-[800px] bg-background p-4 rounded-xl drop-shadow-xl">
+					<p className="font-light text-neutral-500">{`Slow down a bit, you just hit the rate limit! Bandwidth doesn't grow on trees.`}</p>
 					<p>{`Please try to make your request again in a few moments.`}</p>
 				</div>
 			</div>